Finish porting yattag.py to Rust
yattag.py is now just a stub.
